Creates folders for topics
