+++ src/config/ch.sbb.casa/index.js
@@ -1,0 +1,56 @@
+import TrafimageMapboxLayer from '../../layers/TrafimageMapboxLayer';
+import StationsLayer from '../../layers/StationsLayer';
+import {
+  dataLayer,
+  netzkarteLayer,
+  swisstopoSwissImage,
+} from '../ch.sbb.netzkarte';
+
+export const casaDataLayerWithoutLabels = dataLayer.clone({
+  filters: [
+    {
+      field: 'type',
+      value: /symbol|circle/,
+      include: false,
+    },
+  ],
+});
+
+export const casaNetzkarteLayerWithoutLabels = netzkarteLayer.clone({
+  mapboxLayer: casaDataLayerWithoutLabels,
+});
+
+const casaSwisstopoSwissImage = swisstopoSwissImage.clone({
+  mapboxLayer: casaDataLayerWithoutLabels,
+});
+
+export const casaNetzkarteLayerWithLabels = new TrafimageMapboxLayer({
+  name: 'ch.sbb.netzkarte.labels',
+  visible: true,
+  style: 'base_bright_v2',
+  filters: [
+    {
+      field: 'type',
+      value: /symbol|circle/i,
+      include: true,
+    },
+  ],
+  zIndex: 2,
+  properties: {
+    hideInLegend: true,
+  },
+});
+
+// Add stations (blue style on hover) to labelsDataLayer.
+const casaNetzkarteStationsLayer = new StationsLayer({
+  name: 'ch.sbb.netzkarte.stationen.casa',
+  mapboxLayer: casaNetzkarteLayerWithLabels,
+});
+
+export default [
+  casaDataLayerWithoutLabels,
+  casaNetzkarteLayerWithoutLabels,
+  casaSwisstopoSwissImage,
+  casaNetzkarteLayerWithLabels,
+  casaNetzkarteStationsLayer,
+];
